# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Shopping
  description: 在线商城服务
  version: "1.0"
host: localhost:8080
consumes:
- application/json
produces:
- application/json
paths:
  /get_user_info:
    post:
      tags:
      - auth
      summary: 获取用户信息
      description: |-
        获取用户信息

        **Required security scopes for jwt**:
          * `role:user`
      operationId: auth#GetUserInfo
      produces:
      - application/json
      parameters:
      - name: Authorization
        in: header
        description: jwt token
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/AuthGetUserInfoResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
...
  /request_email_code:
    post:
      tags:
      - auth
      summary: 请求电子邮箱验证码
      description: 请求电子邮箱验证码
      operationId: auth#RequestEmailCode
      produces:
      - application/json
      parameters:
      - name: RequestEmailCodeRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/AuthRequestEmailCodeRequestBody'
          required:
          - email
          - action
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/AuthRequestEmailCodeResponseBody'
      schemes:
      - http
  /request_sms_code:
    post:
      tags:
      - auth
      summary: 请求短信验证码
      description: 请求短信验证码
      operationId: auth#RequestSmSCode
      produces:
      - application/json
      parameters:
      - name: RequestSmSCodeRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/AuthRequestSmSCodeRequestBody'
          required:
          - mobile
          - action
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/AuthRequestSmSCodeResponseBody'
      schemes:
      - http
  /reset_password:
    post:
      tags:
      - auth
      summary: 重置密码
      description: 重置密码
      operationId: auth#SetPassword
      produces:
      - application/json
      parameters:
      - name: SetPasswordRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/AuthSetPasswordRequestBody'
          required:
          - username
          - new_password
          - verify_code
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/AuthSetPasswordResponseBody'
      schemes:
      - http
...
definitions:
  AuthGetUserInfoResponseBody:
    title: 'Mediatype identifier: userinfo; view=default'
    type: object
    properties:
      data:
        $ref: '#/definitions/UserResponseBodySimple'
      errcode:
        type: integer
        description: '错误码(0: 成功; 大于0: 失败)'
        example: 0
        minimum: 0
        maximum: 999999
      errmsg:
        type: string
        description: 错误消息
        example: ""
    description: GetUserInfoResponseBody result type (default view)
    example:
      data:
        email: Aut qui.
        id: Aut sit voluptatibus nulla.
        isActive: true
        mobile: Molestiae autem sit odit assumenda.
        nickname: Facilis iste voluptas.
        type: 4978511661108952850
      errcode: 0
      errmsg: ""
    required:
    - errcode
    - errmsg
...
  AuthRequestEmailCodeRequestBody:
    title: AuthRequestEmailCodeRequestBody
    type: object
    properties:
      action:
        type: string
        description: 操作
        example: register
        enum:
        - register
        - reset_password
        - update_email
      email:
        type: string
        description: 电子邮箱
        example: user@example.com
        format: email
    example:
      action: register
      email: user@example.com
    required:
    - email
    - action
  AuthRequestEmailCodeResponseBody:
    title: 'Mediatype identifier: application/vnd.success_result; view=default'
    type: object
    properties:
      ok:
        type: boolean
        description: success
        example: true
    description: RequestEmailCodeResponseBody result type (default view)
    example:
      ok: true
    required:
    - ok
  AuthRequestSmSCodeRequestBody:
    title: AuthRequestSmSCodeRequestBody
    type: object
    properties:
      action:
        type: string
        description: 操作
        example: register
        enum:
        - register
        - reset_password
        - update_mobile
      mobile:
        type: string
        description: 手机号码
        example: "13838003800"
        minLength: 11
        maxLength: 11
    example:
      action: register
      mobile: "13838003800"
    required:
    - mobile
    - action
  AuthRequestSmSCodeResponseBody:
    title: 'Mediatype identifier: application/vnd.success_result; view=default'
    type: object
    properties:
      ok:
        type: boolean
        description: success
        example: true
    description: RequestSmSCodeResponseBody result type (default view)
    example:
      ok: true
    required:
    - ok
  AuthSetPasswordRequestBody:
    title: AuthSetPasswordRequestBody
    type: object
    properties:
      new_password:
        type: string
        description: 新密码
        example: abc123455
        minLength: 6
      username:
        type: string
        description: 用户名（邮箱/手机号）
        example: "13839003900"
        maxLength: 128
      verify_code:
        type: string
        description: 验证码
        example: "888888"
    example:
      new_password: abc123455
      username: "13839003900"
      verify_code: "888888"
    required:
    - username
    - new_password
    - verify_code
  AuthSetPasswordResponseBody:
    title: 'Mediatype identifier: application/vnd.success_result; view=default'
    type: object
    properties:
      ok:
        type: boolean
        description: success
        example: true
    description: SetPasswordResponseBody result type (default view)
    example:
      ok: true
    required:
    - ok
...
  UserResponseBodySimple:
    title: 'Mediatype identifier: user; view=default'
    type: object
    properties:
      email:
        type: string
        description: 邮箱
        example: Esse mollitia maiores et pariatur esse.
      id:
        type: string
        description: ID
        example: Tempora non a.
      isActive:
        type: boolean
        description: 是否可用
        example: false
      mobile:
        type: string
        description: 手机号
        example: Nulla et saepe similique atque.
      nickname:
        type: string
        description: 昵称
        example: Nulla qui distinctio corrupti eveniet aut.
      type:
        type: integer
        description: 类型：1 用户， 2 管理员
        example: 6343364308535236732
        format: int64
    description: 用户 (default view)
    example:
      email: Omnis iure.
      id: Dignissimos ut mollitia et fugit eum amet.
      isActive: false
      mobile: Molestiae et ab.
      nickname: Recusandae nam libero.
      type: 3488167164065706687
    required:
    - id
    - nickname
    - mobile
    - email
    - type
    - isActive
securityDefinitions:
  jwt_header_Authorization:
    type: apiKey
    description: |-
      JWT 认证

      **Security Scopes**:
        * `role:user`: 普通用户
        * `role:admin`: 管理员
    name: Authorization
    in: header